feat: Add PostgresActionRepository and PostgresCommandRepository
This commit adds the `PostgresActionRepository` and `PostgresCommandRepository` structs to the `db` module. These repositories implement the `ActionRepository` and `CommandRepository` traits, respectively, and provide methods for interacting with the PostgreSQL database. The `PostgresActionRepository` includes the `create` and `find_by_pipeline_id` methods, while the `PostgresCommandRepository` includes the `find_by_action_id` and `create` methods. These repositories enable the storage and retrieval of actions and commands in the database.

Recent user commits:
- Refactor imports in pipeline_handler.rs for improved readability
- Refactor grpc_client.rs and grpc_scheduler_client.rs imports
- chore: add libs
- feat: Add find_by_pipeline_id and create methods to ActionRepository
- feat: Add display and conversion implementations for ActionType and ActionStatus
- feat: Add action and command service modules
- feat: Add create_manifest_pipeline method to PipelineService
- feat: Add ActionService trait for handling actions
- feat: Add CommandService trait for handling commands
- feat: Add "name" field to pipelines table